102-fibonacci.c added

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,11 +9,11 @@
 int main(void)
 {
 	int n;
-	int a = 1;
-	int b = 2;
-	int c;
+	long int a = 1;
+	long int b = 2;
+	long int c;
 
-	printf("%d, %d ,", a, b);
+	printf("%ld, %ld ,", a, b);
 
 	for (n = 0; n <= 47; n++)
 	{
@@ -21,12 +21,12 @@ int main(void)
 
 		if (n != 47)
 		{
-			printf("%d, ", c);
+			printf("%ld, ", c);
 			a = b;
 			b = c;
 		}
 		else
-			printf("%d", c);
+			printf("%ld", c);
 	}
 
 	printf("\n");
